requirements.txt
africastalking==1.2.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
contextlib2==21.6.0
distlib==0.3.3
Django==3.2.5
djangorestframework==3.13.1
filelock==3.4.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
platformdirs==2.4.0
pytz==2021.1
requests==2.26.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.7
virtualenv==20.10.0
zipp==3.6.0
